Fix scheduler to return result of main fiber.
The scheduler allocates a new reference to hold the eventual result. If
there is no next fiber in the run queue, then this reference is checked
to see if the main fiber has already exited. If it's not then the main
fiber must be blocked on something (maybe even diverged), so the
scheduler busy loops on checking if something is in the run queue.
When the fiber unblocks at some point, it will be placed into the run
queue and the scheduler can continue.